Refactoring, ComputeHashCallable extracted.
